Modificati test per rispettare le specifiche pre-esistenti

+++ src/main/java/it/unisa/c07/biblionet/gestioneclubdellibro/repository/PostDAO.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PostDAO extends JpaRepository<Post, Long> {
+public interface PostDAO extends JpaRepository<Post, Integer> {
 
-    List<Post> findPostByClubDelLibro_IdClub(long idClub);
+    List<Post> findPostByClubDelLibro_IdClub(int idClub);
 }
